Selective merge of 64-bit branch to corelib
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk@1690 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ pwsafe/pwsafe/corelib/ItemData.h
@@ -214,7 +214,7 @@ public:
   BOOL IsURLEmpty() const {return m_URL.IsEmpty();}
   void SerializePlainText(std::vector<char> &v) const;
   bool DeserializePlainText(const std::vector<char> &v);
-  bool SetField(int type, unsigned char *data, size_t len);
+  bool SetField(int type, unsigned char *data, int len);
 
 private:
   CItemField m_Name;
